# specs/001-oci-registry-server/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: jreg - OCI Distribution Spec v1.1.1
  description: |
    OCI-compliant container registry server implementing the OCI Distribution Specification v1.1.1.
    
    This API enables pushing and pulling OCI images and artifacts using Docker CLI, ORAS, and other OCI-compliant tools.
    
    **Key Features:**
    - Full OCI Distribution Spec v1.1.1 compliance
    - Content-addressable storage (SHA-256/SHA-512)
    - Chunked blob uploads with resume support
    - Multi-architecture image support (OCI Image Index)
    - Referrers API for artifact linking (signatures, SBOMs)
    - Tag management and discovery
    
    **Authentication:** Deferred for MVP (all endpoints publicly accessible)
  version: 1.0.0
  contact:
    name: jreg Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v2
    description: Local development server
  - url: https://registry.example.com/v2
    description: Production server
...
paths:
...
  /{name}/manifests/{reference}:
    parameters:
      - $ref: '#/components/parameters/RepositoryName'
      - $ref: '#/components/parameters/ManifestReference'
    
    head:
      operationId: checkManifestExists
      tags: [manifests]
      summary: Check if manifest exists
      description: |
        Check if a manifest exists (OCI Distribution Spec end-5).
        Reference can be a tag or digest.
      responses:
        '200':
          description: Manifest exists
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            Docker-Content-Digest:
              schema:
                type: string
            Content-Type:
              schema:
                type: string
                enum:
                  - application/vnd.oci.image.manifest.v1+json
                  - application/vnd.oci.image.index.v1+json
                  - application/vnd.docker.distribution.manifest.v2+json
                  - application/vnd.docker.distribution.manifest.list.v2+json
        '404':
          $ref: '#/components/responses/ManifestNotFound'

    # end-6: Pull manifest
    get:
      operationId: getManifest
      tags: [manifests]
      summary: Pull manifest
      description: |
        Download a manifest (OCI Distribution Spec end-6).
        Returns manifest JSON exactly as uploaded (byte-for-byte).
      responses:
        '200':
          description: Manifest content
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            Docker-Content-Digest:
              schema:
                type: string
            Content-Type:
              schema:
                type: string
          content:
            application/vnd.oci.image.manifest.v1+json:
              schema:
                $ref: '#/components/schemas/ImageManifest'
            application/vnd.oci.image.index.v1+json:
              schema:
                $ref: '#/components/schemas/ImageIndex'
        '404':
          $ref: '#/components/responses/ManifestNotFound'

    # end-7: Push manifest
    put:
      operationId: putManifest
      tags: [manifests]
      summary: Push manifest
      description: |
        Upload a manifest (OCI Distribution Spec end-7).
        Associates manifest with a tag if reference is a tag name.
      requestBody:
        description: Manifest JSON
        required: true
        content:
          application/vnd.oci.image.manifest.v1+json:
            schema:
              $ref: '#/components/schemas/ImageManifest'
          application/vnd.oci.image.index.v1+json:
            schema:
              $ref: '#/components/schemas/ImageIndex'
      responses:
        '201':
          description: Manifest uploaded successfully
          headers:
            Location:
              schema:
                type: string
              description: URL to access the manifest
            Docker-Content-Digest:
              schema:
                type: string
        '400':
          description: Invalid manifest JSON or referenced blobs not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    # end-9: Delete manifest
    delete:
      operationId: deleteManifest
      tags: [manifests]
      summary: Delete manifest
      description: |
        Delete a manifest (OCI Distribution Spec end-9).
        If reference is a tag, only the tag is deleted (manifest remains if referenced by other tags or digest).
        If reference is a digest, manifest is deleted (all tags pointing to it become dangling).
      responses:
        '202':
          description: Manifest deletion accepted
        '404':
          $ref: '#/components/responses/ManifestNotFound'

  # end-8a: List tags
  /{name}/tags/list:
    parameters:
      - $ref: '#/components/parameters/RepositoryName'
    
    get:
      operationId: listTags
      tags: [tags]
      summary: List repository tags
      description: |
        List all tags in a repository (OCI Distribution Spec end-8a).
        Supports pagination via `n` and `last` query parameters.
      parameters:
        - name: n
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 10000
          description: Maximum number of tags to return
        - name: last
          in: query
          required: false
          schema:
            type: string
          description: Last tag from previous page (for pagination)
      responses:
        '200':
          description: Tag list
          headers:
            Link:
              schema:
                type: string
                example: </v2/library/ubuntu/tags/list?n=100&last=v1.0.0>; rel="next"
              description: Pagination link (if more results available)
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: Repository name
                    example: library/ubuntu
                  tags:
                    type: array
                    items:
                      type: string
                    description: List of tag names (lexically sorted)
                    example: ["latest", "v1.0.0", "v1.1.0"]
        '404':
          description: Repository not found or has no tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # end-12: List referrers
  /{name}/referrers/{digest}:
    parameters:
      - $ref: '#/components/parameters/RepositoryName'
      - $ref: '#/components/parameters/Digest'
    
    get:
      operationId: listReferrers
      tags: [manifests]
      summary: List referrers
      description: |
        List manifests that reference the specified manifest via `subject` field (OCI Distribution Spec end-12).
        Returns an OCI Image Index containing referrer descriptors.
        Supports filtering by artifact type.
      parameters:
        - name: artifactType
          in: query
          required: false
          schema:
            type: string
            example: application/vnd.example.signature.v1
          description: Filter referrers by artifact type
      responses:
        '200':
          description: Referrers index (may be empty)
          headers:
            Content-Type:
              schema:
                type: string
                example: application/vnd.oci.image.index.v1+json
          content:
            application/vnd.oci.image.index.v1+json:
              schema:
                $ref: '#/components/schemas/ImageIndex'
        '404':
          description: Subject manifest not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    RepositoryName:
      name: name
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9]+((\.|_|__|-+)[a-z0-9]+)*(\/[a-z0-9]+((\.|_|__|-+)[a-z0-9]+)*)*$'
        example: library/ubuntu
      description: Repository name (lowercase, alphanumeric with separators)
    Digest:
      name: digest
      in: path
      required: true
      schema:
        type: string
        pattern: '^(sha256|sha512):[a-f0-9]{64,128}$'
        example: sha256:abc123def456...
      description: Content digest (algorithm:hex)
    ManifestReference:
      name: reference
      in: path
      required: true
      schema:
        type: string
        example: latest
      description: Tag name or digest
...
  schemas:
    ImageManifest:
      type: object
      required:
        - schemaVersion
        - mediaType
        - config
        - layers
      properties:
        schemaVersion:
          type: integer
          enum: [2]
          description: OCI schema version
        mediaType:
          type: string
          enum:
            - application/vnd.oci.image.manifest.v1+json
          description: Manifest media type
        config:
          $ref: '#/components/schemas/Descriptor'
        layers:
          type: array
          items:
            $ref: '#/components/schemas/Descriptor'
        subject:
          $ref: '#/components/schemas/Descriptor'
        annotations:
          type: object
          additionalProperties:
            type: string
      example:
        schemaVersion: 2
        mediaType: application/vnd.oci.image.manifest.v1+json
        config:
          mediaType: application/vnd.oci.image.config.v1+json
          size: 1234
          digest: sha256:abc123...
        layers:
          - mediaType: application/vnd.oci.image.layer.v1.tar+gzip
            size: 5678
            digest: sha256:def456...
    ImageIndex:
      type: object
      required:
        - schemaVersion
        - mediaType
        - manifests
      properties:
        schemaVersion:
          type: integer
          enum: [2]
        mediaType:
          type: string
          enum:
            - application/vnd.oci.image.index.v1+json
        manifests:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Descriptor'
              - type: object
                properties:
                  platform:
                    $ref: '#/components/schemas/Platform'
                  artifactType:
                    type: string
        subject:
          $ref: '#/components/schemas/Descriptor'
        annotations:
          type: object
          additionalProperties:
            type: string
      example:
        schemaVersion: 2
        mediaType: application/vnd.oci.image.index.v1+json
        manifests:
          - mediaType: application/vnd.oci.image.manifest.v1+json
            size: 1234
            digest: sha256:abc123...
            platform:
              architecture: amd64
              os: linux
    Descriptor:
      type: object
      required:
        - mediaType
        - size
        - digest
      properties:
        mediaType:
          type: string
          example: application/vnd.oci.image.layer.v1.tar+gzip
        size:
          type: integer
          format: int64
          example: 1048576
        digest:
          type: string
          pattern: '^(sha256|sha512):[a-f0-9]{64,128}$'
          example: sha256:abc123...
        urls:
          type: array
          items:
            type: string
            format: uri
        annotations:
          type: object
          additionalProperties:
            type: string
        data:
          type: string
          format: byte
          description: Embedded data (base64)
    Platform:
      type: object
      required:
        - architecture
        - os
      properties:
        architecture:
          type: string
          enum: [amd64, arm64, arm, ppc64le, s390x, 386, riscv64]
          example: amd64
        os:
          type: string
          enum: [linux, windows, darwin, freebsd]
          example: linux
        os.version:
          type: string
        os.features:
          type: array
          items:
            type: string
        variant:
          type: string
          example: v8
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - BLOB_UNKNOWN
            - BLOB_UPLOAD_INVALID
            - BLOB_UPLOAD_UNKNOWN
            - DIGEST_INVALID
            - MANIFEST_BLOB_UNKNOWN
            - MANIFEST_INVALID
            - MANIFEST_UNKNOWN
            - NAME_INVALID
            - NAME_UNKNOWN
            - SIZE_INVALID
            - UNAUTHORIZED
            - DENIED
            - UNSUPPORTED
          description: OCI error code
        message:
          type: string
          description: Human-readable error message
        detail:
          type: object
          additionalProperties: true
          description: Additional error context
      example:
        code: BLOB_UNKNOWN
        message: Blob not found in repository
        detail:
          digest: sha256:abc123...
          repository: library/ubuntu
  responses:
...
    ManifestNotFound:
      description: Manifest not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - code: MANIFEST_UNKNOWN
                message: Manifest not found
...